refactor: infinite datalist component ts migration
