+++ packages/components/src/components/infinite-data-list/index.ts
@@ -1,4 +1,4 @@
-import InfiniteDataList from './infinite-data-list.jsx';
+import InfiniteDataList from './infinite-data-list';
 import './infinite-data-list.scss';
 
 export default InfiniteDataList;
