+++ packages/components/stories/icon/icons.js
@@ -695,6 +695,7 @@ export const icons =
         'IcRebrandingDerivGoDashboard',
         'IcRebrandingDerivTraderDashboard',
         'IcRebrandingDerivTrader',
+        'IcRebrandingDerivX',
         'IcRebrandingDerivxWordmark',
         'IcRebrandingDerivx',
         'IcRebrandingDmt5Dashboard',
